+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/ViewModels/AbstractDiaryVM.cs
@@ -16,8 +16,8 @@ namespace TD.WPF.ViewModels
 
             CurrencyRepository = new CurrencyRepository(databasePath);
             SymbolRepository = new SymbolRepository(databasePath, currencyRepository: CurrencyRepository);
-            ExchangeRepository = new ExchangeRepository(databasePath, CurrencyRepository, SymbolRepository);
-            TradingAccountRepository = new TradingAccountRepository(databasePath, ExchangeRepository);
+            ExchangeRepository = new ExchangeRepository(databasePath, currencyRepository: CurrencyRepository, symbolRepository: SymbolRepository);
+            TradingAccountRepository = new TradingAccountRepository(databasePath, exchangeRepository: ExchangeRepository);
 
             Currencies = new ObservableCollection<Currency>();
             Exchanges = new ObservableCollection<Exchange>();